pyproject.toml
[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
force_single_line = true
import_heading_stdlib      = "Standard Library"
import_heading_thirdparty  = "Third Party Library"
import_heading_firstparty  = "First Party Library"
import_heading_localfolder = "Local Library"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

# [tool.mypy-hoge.*]
# ignore_missing_imports = True

[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    "-rxX",
    "--capture=no",
    "-pno:logging",
    # "--cov",
    # "--cov-append",
    # "--cov-report=term-missing",
]
testpaths = "tests"
python_files = [
    "tests.py",
    "test_*.py",
    "*_tests.py",
]


[tool.poetry]
name = "gameserver"
version = "0.1.0"
description = ""
authors = ["pollenjp <polleninjp@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.70.1"
uvicorn = {extras = ["standard"], version = "^0.16.0"}
SQLAlchemy = "^1.4.29"
requests = "^2.26.0"
mysqlclient = "^2.1.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
flake8 = "^4.0.1"
isort = "^5.10.1"
mypy = "^0.930"
pytest = "^6.2.5"
nox = "^2021.10.1"
autoflake8 = "^0.2.2"
sqlalchemy-stubs = "^0.4"
types-PyYAML = "^6.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
